(android) Show full screen qr code when clicked
